Reorganize app Module
 * Reorganize the functions in the app module

 * Move parse_song_string function to bakcend.parsing module

 * Added organization to backend.parsing module
